Rename "sort" to "imports"
